feat: add Enhance project generator and template factory
The diff includes the addition of "EnhanceCliProjectGenerator" and "EnhanceCliProjectModuleBuilder" classes which provide necessary methods to generate an Enhance CLI project. Also, "EnhanceProjectTemplateFactory" class was created to generate the project template via Wizard. Necessary configurations to include these new features into the plugin were made in plugin.xml file. Lastly, a new icon for the project generator was added. These changes aim to facilitate the project generation process for the users of the Enhance CLI.