ci: add usage guard to the build workflow
A new job named "Usage guard" has been added into the GitHub Actions workflow. This job is responsible for running the action-usage-guard@nekofar action from the development branch. This will be useful for authenticating the code and defining the usage guard threshold. This change ensures that the build process does not proceed if guard thresholds are not met, enhancing security.