ci: switch to a GitHub Action for creating release drafts
This commit migrates the manual creation of GitHub release drafts to using the 'nekofar/create-github-release' GitHub action. The change was made to streamline and simplify the process by reducing the need for manual intervention and improving code readability. The functionality remains the same: new releases are still drafted and not publicly visible until manual acceptance, while details of the release such as the tag, title and notes are outlined as before. The use of the action also introduces the feature of marking a release as a pre-release based on certain conditions in the version output.